Enable FF_USE_FASTZIP in MR pipelines
This enables the FF_USE_FASTZIP runner feature flag to use the more
performant compression library 'fastzip' within the runner.

This is promising because we spend a significant amount of time waiting
on artifacts and caches to be (de-)compressed (and up-/downloaded). Any
improvement in performance here will compound across all jobs we run.