Add attach vulnerabilities to new issue action
This adds a new bulk action to the project vulnerability report which
allows you to attach selected vulnerabilities to a newly created issue
and will navigate to that new issue. This is behind the
new_issue_attachment_from_vulnerability_bulk_action feature flag.

The selection summary spec was also slightly reworked.