Add relative root in work_items page query for can_create_branch
When loading a work_item url, the header uses the
"can_create_branch" call to determine if a branch or merge
request can be created for the issue. This call did not
respect the relative_url_root setting.  This lead to this call
returning a 404 and creating unusable elements on the page.

See GitLab issue here:
  https://gitlab.com/gitlab-org/gitlab/-/issues/520134