Fix check for cluster existence in QA cng tool
This change set fixes a subtle bug in the `cng` tool used in QA to check
if the `gitlab` cluster already exists. The problem was that the
`kind get clusters` output actually prints a single existing cluster
name on each line. We want to exactly match a line with the cluster we
are checking for and not just that this cluster name exists anywhere in
the output.