Add minimum override role to CI/CD settings mutation
This adds `ci_pipeline_variables_minimum_override_role` to the
`ProjectCiCdSettingsUpdate` mutation, in preparation for adding this to
the project settings.

It also fixes a bug where the update was performed directly on the
active record object instead of going through `Projects::UpdateService`.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/502258/

Changelog: changed