Add squash_read_only to MergeRequestType
This field is going to be used to determine whether or not squashing
can be set/unset on a merge request

Currently the front end is checking a similar field on ProjectType.
Now that we can set squash requirements based on protected branches
we need to be able to consider the squash setting for a given merge
request.

Contributes to: https://gitlab.com/gitlab-org/gitlab/-/issues/520151
Changelog: added
EE: true