Remove project abilities enabled when user can read_admin_cicd
Users with read_admin_cicd ability should only be able to see minimal
information related to a project and its sub-resources and the
visibility should also be limited to the admin CI/CD pages.
This means, for example, users with read_admin_cicd should
only be able to see a project's path (no URL) and owner name in the
admin CI/CD pages.