Ensure migration names are unique after truncate
In
https://ops.gitlab.net/gitlab-com/database-team/gitlab-com-database-testing/-/jobs/17362259
we can see that after truncating our migration names for Housekeeper we
are getting duplicate file names. This is because we have the following
long migration names that share the same prefix:

```
BackfillProtectedEnvironmentApprovalRulesProtectedEnvironmentProjectId
BackfillProtectedEnvironmentApprovalRulesProtectedEnvironmentGroupId
```

In order to work around this problem we should ensure that our truncated
names always end up unique. We do this by taking a hash of the file
name before truncating and appending this 5 digit number to the end of
the file name. This should be a consistent hash so that each time we run
Housekeeper we get the same name.