Add new policy abilities for path locks
Introduces read_path_locks and create_path_locks to the project policy
and destroy_path_lock to the path lock policy.

Note: admin_path_locks is defined in the project policy and the
path_lock policy. Currently it is exposed in the graph API:

https://gitlab.com/gitlab-org/gitlab/-/blob/master/ee/app/graphql/ee/types/permission_types/project.rb

and also referenced in the path lock helper when checking if a user can
unlock a file:

https://gitlab.com/gitlab-org/gitlab/-/blob/master/ee/app/helpers/path_locks_helper.rb

We will retain the project policy ability to retain backwards
compatability in the API, however, in a future commit we will remove the
path lock policy ability and update the helper to reference
destroy_path_lock which is more descriptive.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/512677