Update the event action field for duo feedback
This commit updates the `EVENT_ACTION` field that goes into the
`duoUserFeedbackMutation` GraphQL mutation.

The app.vue file is the one that actually changes the input to the
mutation. However I did include a change for the user_feedback.vue file
as well even though that is not being used at the moment.

EE: true