Set default for pipeline dependency export
The pipeline dependency export API does not set the default
`export_type` if the parameter is omitted. This means that the record
gets saved with a nil value which causes it to use the database default
of `dependency_list`. This export type is incompatible with the pipeline
export.

Changelog: fixed
EE: true