Fix pre-populated branch name when delete/replace file in overflow menu
Currently, deleting or replacing file when user has permission to push
to branch does not show the correct pre-populated branch name. This
is because we're passing down `targetBranch` from blob_button_group.vue
but the correct variable name is `selectedBranch`.

This feature is behind the `blob_overflow_menu` feature flag.