Fix emoji award button styling for selected state
A recent change in gitlab/ui added specific styling for the selected
state for buttons, which disrupts the styles for emoji award buttons
when the current user has left that emoji award.

This MR adds excplicit styles for the award button selected state so
that they are not overridden.

Changelog: fixed