Fix file breadcrumb link not working in blob view
Currently in the new breadcrumb, we're only passing down `to`,
which is `null` in the blob view. This results in
clicking the link does nothing.

Additionally, this fixes an issue which the new breadcrumb is not
being rendered in blob view since the feature flag is not being
pushed from blob_controller.rb.

This feature is behind the `directory_code_dropdown_updates` flag.