Fetch details of banned and blocked users
This change enables the UpdateSourceUsersService class
to handle cases where source users are missing from the
graphql users response because they are banned or blocked.
The service now collects the ids that are missing from the
response and uses the single user graphql query to
get the user details.

Changelog: changed