Delete the `zoekt_scheduling_worker` feature flag
This feature flag was introduced in 16.9, which is more than 4 milestones ago.

As part of our process we want to ensure [feature flags don't stay too long in the codebase](https://docs.gitlab.com/ee/development/feature_flags/#types-of-feature-flags).

Rollout issue: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/17374

<details><summary>Remaining mentions of the feature flag (click to expand)</summary>

```
config/initializers/1_settings.rb
842:  Settings.cron_jobs['search_zoekt_scheduling_worker'] ||= {}
843:  Settings.cron_jobs['search_zoekt_scheduling_worker']['cron'] ||= '*/1 * * * *'
844:  Settings.cron_jobs['search_zoekt_scheduling_worker']['job_class'] ||= 'Search::Zoekt::SchedulingWorker'

ee/spec/workers/search/zoekt/scheduling_worker_spec.rb
35:          stub_feature_flags(zoekt_scheduling_worker: true)

```

</details>

It is possible that this MR will still need some changes to remove references to the feature flag in the code.
At the moment the `gitlab-housekeeper` is not always capable of removing all references so you must check the diff and pipeline failures to confirm if there are any issues.
It is the responsibility of ~"group::global search" to push those changes to this branch.
If they are already removing this feature flag in another merge request then they can just close this merge request.

## TODO for the reviewers before merging this MR
- [ ] See the status of the rollout by checking https://gitlab.com/gitlab-com/gl-infra/production/-/issues/17374, https://gitlab.com/gitlab-com/gl-infra/feature-flag-log/-/issues/?search=zoekt_scheduling_worker&sort=created_date&state=all&label_name%5B%5D=host%3A%3Agitlab.com and also verify via chatops.
- [ ] [Search for references to `ZoektSchedulingWorker` in frontnd part of code](https://gitlab.com/search?project_id=278964&scope=blobs&search=ZoektSchedulingWorker&regex=false))
- [ ] Check if we need to remove any Gem or other related code by looking at the changes in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/141513

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::DeleteOldFeatureFlags keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: removed