Show previous page if last page becomes empty
This is an improved version of the initial bugfix from
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/182434

Instead of simply jumping to page 1 ("first 20 before null"), we switch
to "last 20 after null", which is effectively equal to "previous page".

And in case the user then navigates "Back" until `hasPreviousPage` turns
false, but that "first page" doesn't have 20 elements, we reload again,
using "first 20 before null" again.