Include predefined_project_variables in workload variables
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/176742 we
introduced the ability to trigger Duo Workflow workflows as CI
pipelines. As part of that work we tried to narrow down the variables
being sent to the CI Job to narrow the scope for security issues and
just generally reduce the surprises for Duo Workflow. In order to
accomplish that we introduced the concept of a "Workload" which, for
now, is just a CI pipeline with way less features than a normal CI
pipeline.

When we shipped this to production we discovered
(https://gitlab.com/gitlab-org/gitlab/-/issues/520931) that we aren't
able to clone the repository. That's because in production we configure
our runners with a custom
[`clone_url`](https://docs.gitlab.com/runner/configuration/advanced-configuration/#how-clone_url-works).

But as we can see from
[the runner code](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/c0ef16085e721ac8074d617004f4eded1a95c730/common/build_url_helper.go#L40)
the runner needs the `$CI_PROJECT_PATH` variable to reconstruct the full
URL to clone the project. So this explains why were seeing cloning
errors where the URL being cloned did not include any project path.

This MR fixes that by passing the full list of
`predefined_project_variables` which includes `$CI_PROJECT_PATH`.

There is a followup action item in
https://gitlab.com/gitlab-org/gitlab/-/issues/517411 to refactor this
allow/deny list of variables that are needed by workloads. Right now
this code is a little bit like shotgun surgery but a larger fix needs to
deal with the fact that right now the responsibility of constructing the
variables is spread across a bunch of different files and such a
refactor is likely to require a lot more investigation.