Documentation: Clarify CI_MERGE_REQUEST_* variable availablity
These variables are available for include:rules but are also exposed
to jobs.

Firstly make this clear in the introduction; the include rules don't
just modify which files are used so clear that up, and link to the
rules to be a bit clearer.

Some of the variables are _not_ set in jobs.  This can be a problem if
you run with "set -u" or similar, so it's worth calling out (my
incorrect assumption was they'd be blank).  I've taken a look at
spec/lib/gitlab/ci/variables/builder/pipeline_spec.rb and marked those
variables that are explicitly tested for absence and marked them with
a note that they are not available under some circumstances.

Co-authored-by: Marcel Amirault <mamirault@gitlab.com>