Treat repeated reassignment to same user as match
This MR updates the bulk reassignment feature to treat a request for an
already pending assignment as a successful match. This is to avoid a
situation where a job restart triggers the same assignment twice,
succeeding the first time and failing the second BECAUSE of the first
round's success.

Related issue: https://gitlab.com/gitlab-org/gitlab/-/issues/519531