Fix invalid sort in placeholder users app
When navigating from the "Members added" tile on group root,
the URL includes "sort=last_joined" parameter. When this is then
passed to the GraphQL query, we get an error due to invalid value
of the `sort` param. This change fixes this bug and only allows
sort values that are available for placeholder users.

Changelog: fixed