Fix invocations of create_current_license in specs
Sometimes we invoke create_current_license with the restrictions
parameter so that we can pass a number of seats for the license.
However, passing the restrictions parameter clobbers the restrictions
in the Gitlab::License where the plan is stored.  So trying to pass
both a plan and restrictions ends up ignoring the plan and the license
ends up on the starter plan.
So we fix this by adding a seats parameter to use instead that will
respect the plan.