Fix group filter on work items list
When filtering by group on the work items list, the UI
currently shows all items (epics, issues, etc) on the
selected group as well as its subgroups. This behaviour
is incorrect.

This commit fixes this behaviour so that the UI only shows
items directly associated with the group (epics) and the
group only (not its subgroups)

Changelog: fixed