Increase length of PBKDF2 user password salts
Increase the length of PBKDF2 user password salts to meet the SP800-132
requirements (128 bits/16 bytes).

Rails uses PBKDF2 in FIPS environments and a random Devise token as a
salt. The token is packed as hex before being used as salt in OpenSSL.

This change increases the length of the generated token from 16 bytes
to 64 bytes. As a result, the length of the salt passed to OpenSSL
increases 8 bytes to 32 bytes.

Resources:
- NIST SP800-132: https://csrc.nist.gov/pubs/sp/800/132/final
- Relates https://gitlab.com/gitlab-org/gitlab/-/issues/522981