Allow has many association for status and lifecycle
We decided that statuses don't belong to a lifecycle
but instead are attached to a lifecycle and
a status can be attached to more than one lifecycle.

Because of this we needed to move the
default statuses for open/closed/duplicate
to the lifecycle.

To increase visibility for which status belongs
to which lifecycle we also moved the association to
the lifecycle (status_ids)