Update move work item service to behave closer to move issue service
Update issue service was running within a transaction. Some
associated data e.g. issues, labels, etc was also copied within
the transaction so if copying some of this associated data failed
the entire move issue operation fails.

We'll want to move away from that in move work item implementation,
but for now we are just implementing the parity, so that there is
less unexpected behaviour before we are ready to document the changes
and implement proper handling mechanisms