Update feature name to GitLab Duo Self-Hosted in the UI
As a result of consultation with Product, Brand Strategy, and Marketing, the decision has been
made to change the name of Self-Hosted Models to GitLab Duo Self-Hosted. The decision is
expected to make marketing and documentation simpler, align the name with a deployment designation
and is only a minor variation from the current name which creates continuity with the feature
as it's currently understood by customers and the media.

See https://gitlab.com/gitlab-org/gitlab/-/issues/515917 for more context

Changelog: changed
EE: true