Add performance bar to GraphQL explorer
It is possible to manually copy the correlation ids from GraphQL
requests into another performance bar. But in order to be able to
iterate quickly on GraphQL queries and see their impact, it'd be nice to
be able to enable the performance bar on the GraphQL explorer site.

There are some issues we need to resolve, e.g. style bleed between our
own styles and the styles of the GraphQL explorer. This is solved with
the help of the ShadowDOM.

Changelog: other