Fix path rejection for Dependency path
The prior implementation of the path rejection relied on array
intersection. This fails because the intersection check does not
preserve order. A short path which has the same elements as a long
path, but misses some intermediate values, would be rejected.

This implementation changes to a pure string approach where
we can guarantee that ordering is maintained.

Changelog: fixed
EE: true