Delete the `zoekt_cache_search_responses` feature flag
This feature flag was introduced in 16.7, which is more than 12 milestones ago.

As part of our process we want to ensure [feature flags don't stay too long in the codebase](https://docs.gitlab.com/ee/development/feature_flags/#types-of-feature-flags).

Rollout issue: (missing URL)

 **Currently the feature flag is `enabled` on production**

It is possible that this MR will still need some changes to remove references to the feature flag in the code.
At the moment the `gitlab-housekeeper` is not always capable of removing all references so you must check the diff and pipeline failures to confirm if there are any issues.
It is the responsibility of ~"group::global search" to push those changes to this branch.
If they are already removing this feature flag in another merge request then they can just close this merge request.

- [ ] See the status of the rollout by checking (missing URL), https://gitlab.com/gitlab-com/gl-infra/feature-flag-log/-/issues/?search=zoekt_cache_search_responses&sort=created_date&state=all&label_name%5B%5D=host%3A%3Agitlab.com
- [ ] Verify the feature flag status via chatops by running `/chatops run feature get zoekt_cache_search_responses`.
- [ ] [Search for references to `ZoektCacheSearchResponses` in frontnd part of code](https://gitlab.com/search?project_id=278964&scope=blobs&search=ZoektCacheSearchResponses&regex=false)
- [ ] [Search for references to `zoekt_cache_search_responses` in code](https://gitlab.com/search?project_id=278964&scope=blobs&search=zoekt_cache_search_responses&regex=false)
- [ ] Check if we need to remove any Gem or other related code by looking at the changes in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/137337

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::DeleteOldFeatureFlags keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: removed