Wrap duo_availability getter logic in with_lock block
Race conditions are happening when attempting to read the duo_availabity value immediately after an update is made to its columns in the setter method. When an update is made in the Duo configuration page, an immediate rediret to admin/gitlab_duo happens where duo_availabity is fetched as a prop for the frontend. Using with_lock ensures any pending updates are fully completed before attempting to read the value.