Use the policy check for related epic links API
The readable_by? method is lacking the correct permissions for public
groups on the Issue (WorkItem) model. This change uses the policy
instead of the overriden readable_by? method to get around this bug.

This led to differently returned related links when the feature flag was
enabled.

There is no performance change, since the Epic model has never overriden
the readable_by method and always used the read policy check.