Sort the terraform reports in the MR widget by "changes required"
Currently, if a pipeline produces a large number of reports (even more
than 4-5) it can be quite difficult to spot reports which have changes.
By sorting by the "number" of changes we at least make it a little
easier to see which more like to be interesting and vice-versa.

Additionally:
* Refactored the report loading from the API response a little to avoid
  duplicate parsing and make the later code a little nicer.
* Add a small test (thanks to @anna_vovchenko) to ensure this doesn't get
  regressed.
* Reports with invalid values for create/update/delete will be sorted to
  to the stop, ensuring they're not missed. Presumably a failed report
  is a sign of bad things and wants to at least be seen.

Changelog: changed

Co-authored-by: Anna Vovchenko <avovchenko@gitlab.com>