refactor: Add oauth_login_counter before request phase middleware
The omniauth gem allows us to add custom logic to the different
phases of the OAuth flow.

Currently, the custom logic in the `before_request_phase` increments
the login counter for the user.

In order to better isolate the custom logic, this commit extracts
the logic into a separate class.

Changelog: other