Track ID Token generation for Secrets Manager
We need to track the various metrics related to
feature usage for Secrets Manager. One of the
metrics is the generation of ID token for accessing the secrets manager.

We define the event and required metrics to track
the ID token generation. We trigger the event when
the ID token is generated in the ProjectSecretsManager.

[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/428871)