Add more detail to sort by id instead of created_at docs
These docs explain that it is usually better to sort by id but they
don't fully explain why this is OK. They currently state that these
sorts are identical which a keen observer will often notice is not true
and they may doubt the rest of the explanation.

The key detail here is that sorting by id is "good enough". We have a
tonne of features where we are sorting by `created_at` which is leading
to a
[bunch of wasteful indexes](https://gitlab.com/gitlab-org/gitlab/-/issues/515865#note_2321526096)
and also could be leading to UX bugs where the sort is unstable.

This MR hopes to clarify things a little.