Add ci_pipeline_schedule_inputs table
This table stores the CI pipeline inputs values configured when creating
a pipeline schedule. It belongs to a pipeline schedule, ensures that the
input keys are unique, and encrypts the input values. It is on the
gitlab_ci database schema because pipeline schedules are also stored on
the CI schema.

Changelog: added
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/519960