Add support for organization-wide agent CI authorization
Uses the newly added 'ci_access.instance' entry in the agent
configuration file to allow all projects within an organization
to use the agent CI tunnel.

Unlike project and group authorizations, specifying an organization
is not allowed because this would create a dependency across
organizations. An agent can only authorize the organization it
belongs to.

This functionality is disabled by default, as it is not currently
suitable for GitLab.com. To use this feature, the application setting
must be enabled.

https://gitlab.com/gitlab-org/gitlab/-/issues/357516

Changelog: added