Standardize path and lfs file locking syncing logic
When LFS files are locked, we create a PathLock, similarly, when a path
is locked which is an LFS file, we create an LfsFileLock.

Previously this logic was duplicated in the PathLock controller and the
GraphQL mutation.

This change moves the syncing logic into the services where it belongs.

It also updates the LFS locking and unlocking services to ensure we
don't have circular logic. When a path lock service calls an lfs
service, we tell the lfs service not to create/destroy the path lock
after creating/destroying the lfs file lock. Similarly, the lfs services
do the same in the opposite direction.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/512677