Only load vulnerability related records when actually exporting
Without this change, the `with_findings_scanner_identifiers_and_notes`
scope eager loads all the supporting models for the serialisation when
we are enqueing the export and will not use this data. This also occurs
before the inner `enumerator` is used, meaning the batch size is too
large to actually carry out.

Changelog: fixed
EE: true