Fix blob overflow menu permalink shortcut does not get triggered
This happens when user navigate to the blob page from repository page.
The `Mousetrap.bind()` in mounted() is getting
ovewritten by the shortcuts_blob.js, which we use to trigger the old
permalink button (when flag is off). This adds a feature flag check
in `shortcuts_blob.js.

Additionally, this also adds a new const under keybindings.js that
better reflects the nature of the new permalink (e.g. copy).

This feature was originally added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/181901 and is
behind the `blob_overflow_menu` feature flag.