Clarify what Project Access Tokens can access
In "cannot access resources from another project" the "from" is a
misleading / incorrect use of the word "from". A PrAT cannot access 
resources that _belong to_ another project. Technically, "from" can be
used in that way in English. But it is very easy to misread that as
saying "you cannot use the PrAT _from_ (in) another project to access
the resources that PrAT has access to".

Reordering the sentence, and removing words like to/from/in, make it
clearer what the permitted behaviour is.