Skip saving pipeline config when using inputs
Inputs may contain secrets, so we don't want to save them in the DB
as plain text. It's safe to not save a pipeline-config because we are
currently considering dropping the `p_ci_pipelines_config` table
because it's not used anywhere.
https://gitlab.com/gitlab-org/gitlab/-/issues/520828#note_2364398251

related to https://gitlab.com/gitlab-org/gitlab/-/issues/523238

Changelog: other