Do not allowlist any missing indexes on foreign keys
We recently had an incident
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19474 in
which a foreign key was added without an index. We already have specs
which detect this problem but they have a large allowlist in there of
different foreign keys without indexes. Clearly there was confusion
about what qualifies as an allowlist as this problem was missed by
several reviewers.

Upon closer investigation it turns out this allowlist
actually had several problems:

1. There were multiple entries saying "index will be added async". But
   this was not even correct. In those cases the column didn't even have
   a foreign key when we introduced it and as such the allowlist wasn't
   necessary. Those cases actually did end up having indexes added
   before the foreign key as necessary and yet they were never removed
   from either allowlist.
2. There were other columns in there which listed "composite index" as a
   reason for not indexing the column. But this spec already accounted
   for composite indexes and implemented the logic correctly.
   Unfortunately those other columns did not have a suitable composite
   index and the existing composite index would have been no use.
3. Other cases were referring to columns that weren't foreign keys at
   all or were indexed so the allowlist was not needed.

So this MR attempts to rectify this situation once and for all with the
following changes:

1. Remove the allowlist altogether
2. Add one exception to the rule which is related to CI partitioned
   tables. Presumably this was due to the rollout of partitions and not
   wanting to create new massive indexes for tables with only a couple
   of partitions. We may be able to get rid of this one day if we have
   many partitions and determine that we need the index to be correct
   now.
3. Add missing indexes for the 2 other columns
4. Remove a foreign key from the `users` table that was missing an
   index. If it's missing then we cannot safely be relying on the
   `CASCADE DELETE` anyway and the validation on `UPDATE/INSERT` is
   something we are usually OK to accept losing with loose foreign keys
   these days.
4. Improve the error output by linking to docs which explain why this
   rule exists. This could hopefully help prevent someone from trying to
   work around this rule in future.

One thing that we can't test, however, in RSpec is whether or not
someone adds the foreign key before adding the index, but in the same
MR. This still has the risk of causing a production incident when the
first migration runs. We may want to consider a solution to this later
potentially using danger or some other clever tricks.

Related to
https://gitlab.com/gitlab-com/gl-infra/production-engineering/-/issues/26412