Improve User#verified_emails when emails are preloaded
When `user.emails` are already loaded calling `user.emails.confirmed`
will issue another database query.

This change checks if the relation is loaded and if it is we use ruby to
select the confirmed emails instead of issuing the database query.

This will help remove N+1 issues when emails are preloaded.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261

Changelog: performance