Cleanup CI/spec setup for real AI requests
We've switched to a centralized evaluation setup to perform e2e
validation of our AI features, and have already removed the specs
that were performing real request to LLM providers (see
https://gitlab.com/gitlab-org/gitlab/-/issues/514012). Here we
cleanup the CI setup, spec helpers, and other unused code. We
also have no need to run an AI Gateway or a LiteLLM service anymore,
so we clean that as well.