Require irb for Rails 7.1 upgrade
This change almost brings back the changes that were reverted from
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/183535

We had reverted because we were getting a load order problem when
running it with our `RUBY_VERSION_DEFAULT: "3.2.6"`, which is only
used on the master branch. It looks like:

1. derailed_benchmarks initializes the application.
2. Zeitwerk attempts to eager load all components.
3. It loads ActionMailbox's InboundEmail model.
4. That model uses has_one_attached from ActiveStorage.
5. ActiveRecord doesn't know about the has_one_attached macro yet
   because ActiveStorage hasn't been loaded yet, and it fails.

The subtle difference is that now we only require
irb inside the and Gitlab::Runtime.console?, which evaluates to false
when derailed_benchmarks runs. So we don't load it at all in this
specific case, which was triggered by the memory-on-boot job.