Add allow_agent_to_request_user to Workflow to disable disambiguation
In
[Duo Workflow](https://handbook.gitlab.com/handbook/engineering/architecture/design-documents/duo_workflow/)
we recently
[added a feature](https://gitlab.com/gitlab-org/duo-workflow/duo-workflow-service/-/merge_requests/284)
which allows the
agent to ask the user for "disambiguation" to clarify any instructions
it was given. This doesn't work well with our initial Duo Workflow in
Remote Execution feature (which runs in CI) as there is no UI for the
user to respond to the question. This means that the workflows run in CI
just get stuck.

As part of https://gitlab.com/gitlab-org/gitlab/-/issues/52494 this new
setting will allow us to just disable this aspect of Duo Workflow. There
may be a case where we want to bring it back later when we build the UI
in the main GitLab application but even then there is still likely to be
cases where we just want to give Duo Workflow a task and let it finish
the task and we'll review the work async.

This MR basically just allows us to set a boolean via the API when
creating the workflow and this boolean will be returned as part of the
Workflow payload.

We need a subsequent MR in the Duo Workflow Service to actually read
this value and disable the disambiguation component when this is false.