Use epic creation adapter in more places
We now use the epic adpter also when creating new epics via the API.
This is still behind a feature flag and by default we still use the
legacy creation service.

This also changes the API spec when creating an epic via the REST API,
which didn't check the correct permissions and allowed guest users to
add a sub-epic.

Changelog: changed
EE: true