Retry SyncServiceTokenWorker for up to 12 times
With this change we will retry SyncServiceTokenWorker
for up to 12 times. This will retry it for up to
approximately 17 hours. If it still fails after this,
we will automatically reschedule the job the next day,
as this job runs daily. This is to ensure this jobs runs
correctly. Since it has external dependencies, we want to
retry it often, since there is more chance at failure.

Changelog: changed
EE: true