Create CodeOwner::Errors abstraction
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261

This change:

1. Introduces a new CodeOwners::Errors class which will be used to
   abstract the logic for adding and evaluating errors
2. Updates the CodeOwners::Error class to simplify the logic.
2. Adds specs for the CodeOwners::Error class.
3. Removes the `path` dependency as this is only used in the
   CodeOwners::Loader#track_file_validation method but the path is
   already accessible in the loader directly from the CodeOwners::File
   object

Currently we only add errors from within the file's logic so the
add_errors method is private.

We are abstracting this logic so it's easier to interact with and test,
and removes some complexity from the CodeOwners::File class.

We also need to be able to add errors to the file class from external
classes to be able to abstract the validation processes.

To keep this MR as small and constrained as possible it is only focusing
on changing how we apply errors. In a following MR we will create a
series of classes to validate the code owners are accessible and have
enough permissions to approve MRs, including the users and groups. These
classes will collect all of the invalid owners and then apply the errors
to the file in one loop.