Fix issue when empty plans sorting to of terraform reports widget in MRs
The intention of the prior MR[1] was that these are sorted to the
bottom, but I misunderstood the behaviour of Number(''). It produces 0,
not NaN. Additionally, it turns out that 0 counts as false for the
purposes of (0 || MAX_VALUE) which is obvious in retrospect. So fix
that, too.

Added some more explanatory comments, and an additional test to cover
the cases discovered here.

[1] https://gitlab.com/gitlab-org/gitlab/-/merge_requests/183300

Changelog: fixed