Use UserCounts for Your Work > To-Do List pill
Currently we let the backend expose the to do count on the super sidebar
nav item, while the User Bar uses the `userCounts` from JS, which is
updatable via the client (e.g. synced from other tabs, when opening
them, or mutated by certain actions in the UI).

By aligning the two, we can make sure that the UserBar at the top and
the count in the sidebar are always aligned.

Changelog: changed