Fix bug in Mutation.ciJobTokenScopeRemoveProject
As per our docs, the default value of "direction"
argument in the GQL mutation to remove a project from the JobToken scope
is "inbound" but in the actual code, it was set to "outbound".
"outbound" allowlist is now deprecated and we should use "inbound"
as the default value as correctly mentioned in the docs.

In the frontend code, we are correctly setting the default
value to "inbound". So nothing should change in the frontend.
The bug was only in the GQL API. This change fixes that.

References:
[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/494322)

Changelog: fixed