Fix award emoji policy for group-level work items
Users have implicit access to group level work items (currently only
epics) when they have access to the project.

This fixes a bug where a user was not allowed to add an emoji to an epic
work item, even though they had read access to that epic (and were able
to comment on it).

Changelog: fixed
EE: true