Allow mapping raw values from code owners refs
Currently the `CodeOwners::ReferenceExtractor` only returns a specific
type of value.

1. names map `@user` -> `user`
2. roles map `@@developer` -> `30`
3. emails map `mail@example.com` -> `mail@example.com`
4. references is a combination of all of these values

In the `CodeOwners::SectionParser` we check if the `default_owners` is
invalid by checking if the `default_owners` value produces any
`references` when parsed by the `CodeOwners::ReferenceExtractor`. This
has a flaw that it passes if at least 1 of the references are formatted
correctly. i.e. `[Section name] @valid_owner malformed_owner` will not
produce an error.

To fix this, we can perform a more advanced check by ensuring all of the
default owners produce a reference. To do this we need to be able to get
back the raw values which were passed in.

This change adds 4 new methods

1. `raw_names`
2. `raw_roles`
3. `raw_emails` (just an alias as these are already the raw values)
4. `raw_references` combination of all the above

This allows us to check if the `default_owners.split` includes all the
same owners as `CodeOwners::ReferenceExtractor.raw_references`

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/525881+