Switch to new validation process for CodeOwners when feature is enabled
This is the first MR in a series of MRs to switch to a new validation
process for Code Owners.

This MR introduces a new wip feature flag named
accessible_code_owners_validation which will switch between the old
validation process and the new one.

To keep the scope of this change as focused as possible it is limited to
changes in the CodeOwners::File only.

A following change will add logic to the
CodeOwners::OwnerValidation::Process to validate accessible code owners.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/429261