graphql: Don't get the commit from paginatedTree
The commit information box displayed at the top of the file browser in
the GitLab UI is rendered by calling the `pathLastCommit` GraphQL
operation. This operation makes a query to `paginatedTree` as a way to
retrieve the details of the last commit that was made for a given path.

paginatedTree eventually calls the Gitaly GetTreeEntries RPC, which
expects a path to a "tree-ish" [1] as input.

There are two problems with this approach:

1. When users navigate to a particular file in the UI, the paginatedTree
   query takes the _filename_ as the `path` argument. When Gitaly
   attempts to validate this, it fails because a file is not "tree-ish".
   The failure produces around half a million logs on GitLab.com every
   24 hours as it's captured as a legitimate RPC failure. The GraphQL
   resolver however will ignore this error, as it doesn't actually need
   the tree itself.
2. Calling a Gitaly GetTreeEntries RPC when a tree is not actually
   needed is fairly inefficient. We should call the FindCommits RPC
   directly instead.

This is the first of two commits which should be merged and deployed
sequentially; the second commit should only be deployed once the backend
changes have completely rolled out.

Modify the LastCommitResolver so that explicit arguments can be provided
for the commit lookup -- it's no longer dependent on being nested in a
Tree object. Instead, it's now expected to be nested under a Repository
type, so we also add last_commit as a field to the Repository.

[1] https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-aiddeftree-ishatree-ishalsotreeish